Updated to PHP 7.4 standards
