+++ etc/di.xml
@@ -3,10 +3,10 @@
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
 
     <!-- Console command to clean up bot sessions from db -->
-    <type name="Magento\Framework\Console\CommandList">
+    <type name="Magento\Framework\Console\CommandListInterface">
         <arguments>
             <argument name="commands" xsi:type="array">
-                <item name="fl32_bot_sess_clean" xsi:type="object">Flancer32\BotSess\Cli\Clean</item>
+                <item name="fl32_bot_sess_clean" xsi:type="object">Flancer32\BotSess\Command\Clean</item>
             </argument>
         </arguments>
     </type>
